requirements.txt
google_api_python_client==2.156.0
google_auth_oauthlib==1.2.1
langchain==0.3.13
langchain_community==0.3.13
langchain_openai==0.2.14
protobuf==5.29.2
python-dotenv==1.0.1
